+++ CHANGELOG.md
@@ -15,6 +15,7 @@ IMPROVEMENTS:
 - Validate public key format and type (PEM format, RSA type is allowed).
 - Decrease message payload size when sending over message queue by sending as bytes using Protobuf.
 - Decrease message queue receiving size limit to 3.25MB.
+- Verify accessor signature when IdP sending a response (calling POST `/idp/response`).
 
 BUG FIXES: